Add class for making request in client project, change signature of incoming file store request, change of file encoding.

+++ src/main/java/com/hatexomatulak/snes/repositories/CategoryRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface CategoryRepository extends CrudRepository<CategoryDTO, Long> {
+public interface CategoryRepository extends CrudRepository<CategoryDTO, Integer> {
     List<CategoryDTO> findAll();
 }
